+++ NITC_Hostel_Manager_Frontend/app/src/main/java/com/kumar/nitchostelmanager/profile/fragments/StudentDashboardFragment.kt
@@ -1,9 +1,11 @@
 package com.kumar.nitchostelmanager.profile.fragments
 
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.ViewModel
@@ -68,16 +70,29 @@ class StudentDashboardFragment:Fragment() ,CircleLoadingDialog{
         }
 
         binding.logoutButtonInStudentDashboard.setOnClickListener {
+            showDialog()
+        }
+
+        return binding.root
+    }
+
+    private fun showDialog(){
+        val dialog = activity?.let { AlertDialog.Builder(it) }
+        dialog?.setCancelable(false)
+        dialog?.setTitle("Logout")
+        dialog?.setMessage("Are you sure you want to log out?")
+        dialog?.setNegativeButton("No", DialogInterface.OnClickListener{ dialog, which ->
+            dialog.cancel()
+        })
+        dialog?.setPositiveButton("Yes", DialogInterface.OnClickListener { dialog, which ->
             var deleted = LocalStorageAccess(
                 this@StudentDashboardFragment,
                 requireContext(),
                 profileViewModel
             ).deleteData()
-
             findNavController().navigate(R.id.loginFragment)
-        }
-
-        return binding.root
+        })
+        dialog?.create()?.show()
     }
 
     private fun getStudentProfile(){
